refactor controlnet and add img2img and inpaint
